ASYLUM: Simplify the console code
