+++ engines/asylum/asylum.cpp
@@ -50,12 +50,13 @@
 #include "asylum/views/menu.h"
 #include "asylum/views/video.h"
 
+#include "asylum/console.h"
 #include "asylum/respack.h"
 
 namespace Asylum {
 
 AsylumEngine::AsylumEngine(OSystem *system, const ADGameDescription *gd) : Engine(system), _gameDescription(gd),
-	_console(nullptr), _cursor(nullptr), _encounter(nullptr), _menu(nullptr), _resource(nullptr), _savegame(nullptr),
+	_cursor(nullptr), _encounter(nullptr), _menu(nullptr), _resource(nullptr), _savegame(nullptr),
 	_scene(nullptr), _screen(nullptr), _script(nullptr), _special(nullptr), _speech(nullptr), _sound(nullptr), _text(nullptr),
 	_video(nullptr), _handler(nullptr), _puzzles(nullptr) {
 
@@ -112,8 +113,7 @@ Common::Error AsylumEngine::run() {
 	initGraphics(640, 480);
 
 	// Create debugger. It requires GFX to be initialized
-	_console   = new Console(this);
-	setDebugger(_console);
+	setDebugger(new Console(this));
 
 	// Create resource manager
 	_resource  = new ResourceManager(this);
@@ -392,11 +392,8 @@ void AsylumEngine::playIntro() {
 }
 
 void AsylumEngine::handleEvents() {
-	if (!_console || !_video || !_screen || !_sound || !_menu || !_cursor)
+	if (!_video || !_screen || !_sound || !_menu || !_cursor)
 		error("[AsylumEngine::handleEvents] Subsystems not initialized properly!");
-
-	// Show the debugger if required
-	_console->onFrame();
 
 	AsylumEvent ev;
 	Common::Keymapper *const keymapper = _eventMan->getKeymapper();
@@ -414,11 +411,6 @@ void AsylumEngine::handleEvents() {
 			break;
 
 		case Common::EVENT_KEYDOWN:
-			if ((ev.kbd.flags & Common::KBD_CTRL) && ev.kbd.keycode == Common::KEYCODE_d) {
-				_console->attach();
-				break;
-			}
-
 			// Handle key events
 			if (_handler)
 				_handler->handleEvent(ev);
