+++ engines/asylum/detection.h
@@ -8,65 +8,39 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
-
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
-
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
-#ifndef ASYLUM_PUZZLES_PUZZLES_H
-#define ASYLUM_PUZZLES_PUZZLES_H
-
-#include "common/serializer.h"
-
-#include "asylum/shared.h"
+#ifndef ASYLUM_DETECTION_H
+#define ASYLUM_DETECTION_H
 
 namespace Asylum {
 
-class AsylumEngine;
-class EventHandler;
-class Puzzle;
-
-class Puzzles : public Common::Serializable {
-public:
-	Puzzles(AsylumEngine *engine);
-	~Puzzles();
-
-	/**
-	 * Resets puzzles
-	 */
-	void reset();
-
-	/**
-	 * Gets a message handler.
-	 *
-	 * @param index Zero-based index of the message handler
-	 *
-	 * @return The message handler.
-	 */
-	EventHandler *getPuzzle(uint32 index) const;
-
-	// Serializable
-	void saveLoadWithSerializer(Common::Serializer &s);
-
-private:
-	AsylumEngine *_vm;
-	Puzzle *_puzzles[17];
-
-	/**
-	 * Initializes the puzzles
-	 */
-	void initPuzzles();
-
-	// Debug
-	friend class Console;
+enum kDebugLevels {
+	kDebugLevelMain      = 1 << 0,
+	kDebugLevelResources = 1 << 1,
+	kDebugLevelSprites   = 1 << 2,
+	kDebugLevelInput     = 1 << 3,
+	kDebugLevelMenu      = 1 << 4,
+	kDebugLevelScripts   = 1 << 5,
+	kDebugLevelSound     = 1 << 6,
+	kDebugLevelSavegame  = 1 << 7,
+	kDebugLevelScene     = 1 << 8,
+	kDebugLevelObjects   = 1 << 9,
+	kDebugLevelActor     = 1 << 10,
+	kDebugLevelEncounter = 1 << 11,
+	kDebugLevelVideo     = 1 << 12,
+	kDebugLevelCommands  = 1 << 13
 };
 
 } // End of namespace Asylum
 
-#endif // ASYLUM_PUZZLES_PUZZLES_H
+#endif // ASYLUM_DETECTION_H
